aider: Added console log to print out the data returned from the API request.
# Context:
USER: now im getting this response in the web ui : No videos found for the provided channel ID and search criteria.
ASSISTANT: This message is displayed when the `data.pageInfo.totalResults` is 0 or when the `videos` array is not an array. This could mean that the API request to YouTube is not returning any videos for the provided channel ID and search criteria.

To debug this, we can add some console logs to the `getChannelVideos` function in the `app.js` file to print out the data returned from the API request. Here's how you can modify the code:

```javascript
app.js
<<<<<<< ORIGINAL
.then(data => {
    if (data.pageInfo && data.pageInfo.totalResults === 0) {
        showError("No videos found for the provided channel ID and search criteria.");
    } else {
        displayVideos(data.items);
    }
})
=======
.then(data => {
    console.log(data); // Add this line
    if (data.pageInfo && data.pageInfo.totalResults === 0) {
        showError("No videos found for the provided channel ID and search criteria.");
    } else {
        displayVideos(data.items);
    }
})
>>>>>>> UPDATED
```

This change adds a console log to print out the data returned from the API request. You can check the console in your browser's developer tools to see the output. This should help you understand why no videos are being returned.