aider: Added console log to print out the requested URL.
# Context:
USER: yeah you are incorrect the youtubechannelid is correct!
ASSISTANT: If the YouTube channel ID is correct, then the 404 error might be due to an incorrect API endpoint or a problem with the server.

To debug this, you can add a console log in your `api/youtube-api.js` file to print out the URL being requested. This will help you verify that the URL is correct.

Here's the change:

```javascript
api/youtube-api.js
<<<<<<< ORIGINAL
  const url = `https://www.googleapis.com/youtube/v3/search?key=${apiKey}&channelId=${channelId}&part=snippet,id&order=date&maxResults=${maxResults}${searchQuery ? `&q=${encodeURIComponent(searchQuery)}` : ''}`;

  try {
    const response = await axios.get(url);
    res.status(200).json(response.data);
  } catch (error) {
    res.status(error.response.status).json({ message: error.message });
  }
=======
  const url = `https://www.googleapis.com/youtube/v3/search?key=${apiKey}&channelId=${channelId}&part=snippet,id&order=date&maxResults=${maxResults}${searchQuery ? `&q=${encodeURIComponent(searchQuery)}` : ''}`;

  console.log(`Requesting URL: ${url}`);

  try {
    const response = await axios.get(url);
    res.status(200).json(response.data);
  } catch (error) {
    res.status(error.response.status).json({ message: error.message });
  }
>>>>>>> UPDATED
```

After making this change, you can check the server logs to see the URL being requested. If the URL is correct, the problem might be with the server or the YouTube API.